+++ scripts/check_whitespace.sh
@@ -26,6 +26,8 @@ options=(
 )
 
 git grep "${options[@]}" -- \
+    ':(exclude)Firebase/CoreDiagnostics/FIRCDLibrary/Protogen/nanopb' \
+    ':(exclude)Firebase/CoreDiagnostics/ProtoSupport' \
     ':(exclude)Firestore/Protos/nanopb' \
     ':(exclude)Firestore/Protos/cpp' \
     ':(exclude)Firestore/Protos/objc' \